update: 実装を適切なファイルへ移動
