+++ internal/repository/errors.go
@@ -1,0 +1,10 @@
+package repository
+
+import (
+	"errors"
+)
+
+var (
+	ErrorUnsupportedConfig     = errors.New("unsupported config")
+	ErrorUnsupportedRepository = errors.New("unsupported repository")
+)
